[WIP] Integrate bgfx tools with truss. (#30)
* Added back tool compilation.

This is the first step towards tool integration.

* Added generateParsers step to Linux build process.

On Linux, an extra step is required to build bgfx tools. Flex and
Bison must be used to compile the GLSL-optimizer code.  This commit
adds an additional build step to ExternalProject to handle this.

* Added binary build command for bgfx tools.

This automatically builds and installs the bgfx tools into a `bin`
directory in the truss distribution.

* Refactored .gitignore into separate files.

This creates a separate .gitignore for the truss distribution
directory, which is useful for people who are going to be creating
standalone distributions.

* Added binary auto-extraction to truss core startup.